Add files via upload
Stable well-tested code suitable for long-term recording of sensor data such as seismic data. This be easily adapted to record a range of input signals periodically producing plots for uploading to the web. Data are saved in timestamped .mseed format for ease of analysis by standard seismic software. This code runs for months or years without user intervention and has been used by others for a range of sensing projects.

    Seismic Monitoring Code Description
    rpiseismic.py - runs on Raspberry Pi to records and plots data
    InfraSoundAnalyser.py - outline analysis code for PC